Remove initial code duplication between mssql clients
